# docs2/cmdb-api-spec.yaml
openapi: 3.0.1
info:
  title: Chelas Movies DataBase
  description: "An API that obtains movies from the IMDB API and manages user's and the groups of movies made by the users (their movie lists)"
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
paths:
...
  /login:
    post:
      tags:
        - Login
      summary: Uses authentication credentials to get the user's token
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogIn'
      responses:
        200:
          description: Returns token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTokenResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Password error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'                
...
  /groups/{groupID}:
    get:
      tags:
        - Groups
      summary: Returns detail of a Group
      parameters:
        - name: groupID
          in: path
          description: the Group's ID to return the details of
          required: true
          schema:
            type: integer
            example: 5
      responses:
        200:
          description: Returns a user's group detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMissingParam'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotLogIn'   
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUserNF' 
                
    post:
      tags:
        - Groups
      summary: Updates a Group
      parameters:
        - name: groupID
          in: path
          description: the Group's ID to update
          required: true
          schema:
            type: integer
            example: 5
      requestBody:
        description: Required information for updating a group
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        200:
          description: Returns only the status code

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorGroupNameEmpty'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotLogIn'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUserNF' 
                
    delete:
      tags:
        - Groups
      summary: Deletes a Group
      parameters:
        - name: groupID
          in: path
          description: the group's identification to be deleted
          required: true
          schema:
            type: integer
            example: 5
      responses:
        200:
          description: Response with the information of group's deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteGroupResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMissingGroupID'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotLogIn'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUserNF'               
...
components:
  schemas:
    ErrorUserNF:
      type: object
      description: Application error
      properties:
        error:
          type: string
          description: The returned error
          example: "User not found"
    ErrorNotLogIn:
      type: object
      description: Application error
      properties:
        error:
          type: string
          description: The returned error
          example: "You are not logged in / You have no authorization to perform this action"
...
    ErrorGroupNameEmpty:
      type: object
      description: Application error
      properties:
        error:
          type: string
          description: The returned error
          example: "Group name must be a non-empty string"
    ErrorMissingParam:    
      type: object
      description: Application error
      properties:
        error:
          type: string
          description: The returned error
          example: "Path param {Name} is missing"
    ErrorMissingGroupID:    
      type: object
      description: Application error
      properties:
        error:
          type: string
          description: The returned error
          example: "Group with id {groupID} not found"
...
    AppError:
      type: object
      description: Application error
      required:
        - code
        - name
        - message
      properties:
        code:
          type: number
          format: integer
          description: HTTP Error code
          example: 4**
        name:
          type: string
          description: HTTP Error name
          example: Client Error name
        message:
          type: string
          description: Error description message
          example: An error occurred
...
    UserLogIn:
      description: Information for user authentication
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: the user's name
          example: Filipe
        password:
          type: string
          description: the user's password (which will be hashed)
          example: strongpass
    UpdateGroupRequest:
      description: Update Details of a Group
      required:
         - groupName
         - groupDescription
      properties:
        groupName:
          type: string
          description: the group's name to update
          example: "Last seen"
        groupDescription:
          type: string
          description: the group's description to update
          example: "Of action films"
    UserTokenResponse:
      type: object
      description: Response with the user's id and token
      properties:
        userID:
          type: string
          description: the user token
          example: 2
        token:
          type: string
          description: the user token
          example: "061bebd1-0d4b-43e4-b2d2-34f1be9e0cab"
    DeleteGroupResponse:
      type: object
      description: Group deletion message
      properties:
        msg:
          type: string
          description: Message 
          example: "Deleted group w/ id -> {groupID} from user -> {userName} "
...
    GroupDetailResponse:
      type: object
      description: Group detail
      required:
        - id
        - name
        - description
        - isPrivate
        - movies
        - totalDuration
      properties:
        id:
          type: string
          format: integer
          description: Group's ID
          example: 1
        groupObj:
          type: object
          description: Movies information
          required:
            - name
            - description
            - isPrivate
            - movies
            - totalDuration
          properties:
            name:
              type: string
              description: Group's name
              example: watch later
            description:
              type: string
              description: Group's description
              example: No time
            isPrivate:
              type: boolean
              description: Group's privacy
              example: true
            movies:
              type: array
              items:
                type: object
                description: Movies information
                required:
                  - id
                  - name
                  - duration
                properties:
                  id:
                    type: string
                    description: The movie's ID
                    example: tt0120663        
                  name:
                    type: string
                    description: The movie's name
                    example: Eyes Wide Shut                            
                  duration:
                    type: number
                    format: integer
                    description: Duration (min) of the movie
                    example: 159
            totalDuration:
              type: number
              format: integer
              description: Total duration (min) of movies in group
              example: 159        
...